apps/file-generator/requirements/base.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings>=2.2.1
PyYAML==6.0.2
firebase-admin>=6.5.0
google-api-python-client>=2.124.0
google-auth>=2.29.0
google-cloud-translate>=3.15.0
docxtpl>=0.16.8
